Moved everything to correct locations
